controllers and converters created

+++ src/main/java/pl/shellchuck/charity/repository/DonationRepository.java
@@ -3,5 +3,5 @@ package pl.shellchuck.charity.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.shellchuck.charity.entity.Donation;
 
-public interface DonationRepository extends JpaRepository<Donation, Integer> {
+public interface DonationRepository extends JpaRepository<Donation, Long> {
 }
